+++ .gitignore
@@ -1,1 +1,4 @@
 node_modules
+lighthouse
+playground-1.mongodb.js
+.env
