+++ package.json
@@ -7,6 +7,14 @@
     "start": "node server.improved.js"
   },
   "dependencies": {
-    "mime": "^2.6.0"
+    "cookie-session": "^2.0.0",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "mime": "^2.6.0",
+    "mongodb": "^6.1.0",
+    "sanitize.css": "^13.0.0"
+  },
+  "engines": {
+    "node": "16.x"
   }
 }
