+++ server.improved.js
@@ -1,207 +1,40 @@
 const http = require( 'http' ),
       fs   = require( 'fs' ),
-      // IMPORTANT: you must run `npm install` in the directory for this assignment
-      // to install the mime library if you're testing this on your local machine.
-      // However, Glitch will install it automatically by looking in your package.json
-      // file.
+      express = require('express'),
+      cookie = require('cookie-session'),
+      dotenv = require("dotenv"),
+      { MongoClient, ObjectId } = require("mongodb"),
       mime = require( 'mime' ),
       dir  = 'public/',
-      port = 3000
+      port = 3000,
+      app = express();
 
-const appdata = [
-  { 'Name': 'Webware', 
-  'Code': 'CS4241',
-  'StartTime' : '12:00',
-  'EndTime': '14:00',
-  'Days': { 'pos0': 'M', 'pos1': 'Th' },
-  'Length': 2 },
-]
+// allows use of environment variables
+dotenv.config()
 
-const server = http.createServer( function( request,response ) {
-  if( request.method === 'GET' ) {
-    handleGet( request, response )    
-  }else if( request.method === 'POST' ){
-    handlePost( request, response ) 
-  }
-})
+const handleLogin = function(request, response) {
+  console.log('body: ' + JSON.stringify(request.body))
+  console.log('un: ' + request.body.username + " pw: " +request.body.password)
 
-const handleGet = function( request, response ) {
-  const filename = dir + request.url.slice( 1 ) 
+  // check if the username matches a username in the database
+  if( request.body.username == 'test' ) { // TODO
+    // check if the password matches
+    if(request.body.password == 'test') { // TODO
+      // request.session.login == true
+      response.redirect("https://www.google.com/") // check this out
+      console.log('correct!')
+    } else {
+      // password incorrect, redirect back to login page
+      response.sendFile( __dirname + '/public/index.html' )
 
-  if( request.url === '/' ) {
-    sendFile( response, 'public/index.html' )
-  }else{
-    sendFile( response, filename )
-  }
-}
+      console.log('fail!')
 
-const handlePost = function( request, response ) {
-
-
-  if(request.url === '/submit') {
-    handleSubmit( request, response )
-  } else if(request.url === '/getAll') {    
-    handleGetAll( request, response )
-  } else if(request.url === '/remove') {
-    handleRemove( request, response )
-  } else if(request.url === '/modify') {
-    handleModify( request, response )
-  } else {
-    console.log('FAIL - requested: ' + request.url)
-  }
-}
-
-const handleGetAll = function(request, response) {
-    request.on( 'data', () => {})
-    request.on( 'end', function() {          
-      response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-      // console.log('getAll - sending: ' + JSON.stringify(appdata))
-      response.end(JSON.stringify(appdata))
-    })
-}
-
-const handleSubmit = function(request, response) {
-    let dataString = ''
-
-    request.on( 'data', function( data ) {
-        dataString += data 
-    })
-
-    request.on( 'end', function() {
-      // console.log( 'submit - received: ' + JSON.parse( dataString ) )
-  
-      json = JSON.parse( dataString )
-        
-      // add data
-      // check has name, start time, end time, and 1 day associated
-      const error = validate(json);
-
-      if(error.errors === false) { // no errors
-
-        // calculate the derived field (length of class)
-        json['Length'] = calcDerivedLength(json.StartTime, json.EndTime)
-
-        // add to the server data
-        appdata.push(
-          json
-        )
-
-        // console.log('submit - new data: ' + JSON.stringify(appdata))
-
-        response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-        response.end(JSON.stringify({}))
-      } else { // send back error message
-        // console.log('submit - error: ' + JSON.stringify(error))
-        response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-        response.end( JSON.stringify(error) )
-      }
-    })
-}
-
-const handleRemove = function(request, response) {
-    let dataString = ''
-
-    request.on( 'data', function( data ) {
-        dataString += data 
-    })
-
-    request.on( 'end', function() {
-      // console.log( 'remove - received: ' + JSON.parse( dataString ) )
-  
-      json = JSON.parse( dataString )
-        
-      // find the data to remove
-      let i = 0;
-      const max = appdata.length
-      for(obj of appdata) {
-        if(JSON.stringify(obj) === json) {
-            const front = appdata.slice(0, i)
-            const back = appdata.slice(i+1)
-            const temp = front.concat(back)
-            while(appdata.length > 0) {
-                appdata.pop()
-            }
-            for(let j = 0; j < temp.length; j++) {
-                appdata.push(temp[j])
-            }
-            
-            response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-            response.end('success')
-            console.log('remove - success')
-            break
-        } else {
-            i++
-        }
-      }
-
-    if(i >= max) {
-        response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-        response.end('fail')
-        console.log('failed to remove')
     }
-    })
-}
-
-const handleModify = function(request, response) {
-    let dataString = ''
-
-    request.on( 'data', function( data ) {
-        dataString += data 
-    })
-
-    request.on( 'end', function() {
-      // console.log( 'modify - received: ' +  dataString )
-  
-      json = JSON.parse( dataString )
-        
-      prev = json.prev
-      data = json.new
-
-      if(prev === "") { // no classes to modify
-        const error = {
-          'errors': true,
-          'classModifySelect': "No Classes To Modify",
-        }
-        // console.log('modify - error: ' + JSON.stringify(error))
-        response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-        response.end( JSON.stringify(error) )
-      } else {
-        // validate the new values
-        const error = validate(data);
-
-        if(error.errors === false) {
-          // find the data to modify
-          let i = 0;
-          const max = appdata.length
-          for(obj of appdata) {
-              if(JSON.stringify(obj) === prev) {
-
-                  obj.Name = data.Name
-                  obj.Code = data.Code
-                  obj.StartTime = data.StartTime
-                  obj.EndTime = data.EndTime
-                  obj.Length = calcDerivedLength(data.StartTime, data.EndTime)
-                  
-                  response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-                  response.end(JSON.stringify({}))
-                  console.log('modify - success')
-                  break
-              } else {
-                  i++
-              }
-          }
-          if(i >= max) {
-              response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-              response.end(JSON.stringify(error))
-              console.log('failed to modify')
-          }
-        } else { // send back error message
-        //   console.log('modify - error: ' + JSON.stringify(error))
-          response.writeHead( 200, "OK", {'Content-Type': 'text/plain' })
-          response.end( JSON.stringify(error) )
-        }        
-      }
-    })
+  } else {
+    // password incorrect, redirect back to login page
+    response.sendFile( __dirname + '/public/index.html' )
+    console.log('fail!')
+  }
 }
 
 const sendFile = function( response, filename ) {
@@ -209,7 +42,7 @@ const sendFile = function( response, filename ) {
 
    fs.readFile( filename, function( err, content ) {
 
-     // if the error = null, then we've loaded the file successfully
+    // if the error = null, then we've loaded the file successfully
      if( err === null ) {
 
        // status code: https://httpstatuses.com
@@ -226,50 +59,66 @@ const sendFile = function( response, filename ) {
    })
 }
 
-const validate = data => {
-  error = {
-    errors: false,
-    className: true,
-    startTime: true,
-    endTime: true,
-    days:  true
-  }
-  if(!(data.Name.length > 0) ) {
-    error.className = "Name required"
-    error.errors = true
-  }
-  if(!(data.StartTime.length > 0) ) {
-    error.startTime = "Start Time required"
-    error.errors = true
-  }
-  if(!(data.EndTime.length > 0) ) {
-    error.endTime = "End Time required"
-    error.errors = true
-  }
 
-  const start = data.StartTime.split(':')
-  const end = data.EndTime.split(':')
-  if(start[0] > end[0]) {
-    error.startTime = "Start Time must be before End Time"
-    error.errors = true
-  } else if (start[0] == end[0] && start[1] > end[1]) {
-    error.startTime = "Start Time must be before End Time"
-    error.errors = true
-  }
+// express set up
 
-  if(!(Object.keys(data.Days).length > 0)) {
-    error.days = "Must select at least one day"
-    error.errors = true
-  }
-  return error;
-};
+// cookie set up
 
-const calcDerivedLength = (start, end) => {
-    // calculate the derived field (length of class)
-    const s = new Date('1970-01-01T' + start + ":00")
-    const e = new Date('1970-01-01T' + end + ":00")
-    const msPerHour = 1000 * 60 * 60
-    return Math.round(((e-s) / msPerHour) * 100) / 100
+// use express.urlencoded to get data sent by default form actions
+// or GET requests
+app.use( express.urlencoded({ extended: true }) )
+
+// public directory
+app.use(express.static('./public'))
+app.use(express.json())
+
+let collection = null
+
+// Database Connection
+const uri = `mongodb+srv://${process.env.USER}:${process.env.PASS}@${process.env.HOST}`
+const client = new MongoClient( uri )
+
+async function run() {
+  await client.connect()
+  collection = await client.db("datatest").collection("test")
 }
+run()
 
-server.listen( process.env.PORT || port )
+
+// route to get all docs
+app.get("/docs", async (request, response) => {
+  if (collection !== null) {
+    const docs = await collection.find({}).toArray()
+    response.json( docs )
+  }
+})
+
+app.post( '/add', async (request,response) => {
+  const result = await collection.insertOne( request.body )
+  response.json( result )
+})
+
+app.use( (request,response,next) => {
+  console.log('idk what this function does but it just got called!')
+  if( collection !== null ) {
+    console.log('what does next do?')
+    next()
+  }else{
+    console.log('503?')
+    response.status( 503 ).send()
+  }
+})
+
+// set up GET and POST requests
+
+// GET requests
+app.get('/', function (request, response) {
+  console.log('getting index.html')
+  response.sendFile( __dirname + '/public/index.html' )
+})
+
+// POST requests
+app.post('/login', handleLogin)
+
+// set up the server
+app.listen(3000)
